nfsv4.1: refactor reply cache
the earlier implementation of reply cache was encapsulated
inside session. On one hand it works, but on the other hand
ugly implementation did not allow me to sleep good.

This patch make nfs server session and cache aware.
The CompoundContext knows about number of operations and
current operation position.

The operation position verification is not spec complaint.
(we set compound status but do not return operation result).
Nevertheless it protects against malicious clients.