nfs4: refactor/optimize data structures
The nfs spec has some structures which defined with typedef:

typedef unsigned int   uint32_t;
typedef uint32_t        mode4;

and so on.

The jrpcgen uses aggregation to implement typedef relations. As a result,
in some cases a chain of objects created to store a single primitive value.

This patch addresses this by replacing aggregation with inheritance.
This is first one in this work and touches mostly used types.

Acked-by: Gerd Behrmann
Target: master