nfs: use mutimap to store export entries
The export file is parsed and represented as a list of FsExport recored
sorted by smallest network first order. This list is scanned for any
access by file handle. If we split the list by export index ( export path ),
then we need to scan only a portion of list, which is relevant for
particular request.

Can improve performance on installations with a large number of exports.

Acked-by: Gerd Behrmann
Target: master