nfs4.0: refactor handling of open-sequence id with v4.0
up to now, we was resetting sequence on each open and
did not check it on open_confirm or close.

with this fix, some py-nfs tests started to fail (mostly
retransmit related) as have ignored sequence id and did not
detect retransmits (the existing retransmit detection code
is for v4.1 only).

Acked-by: Gerd Behrmann
Target: master