nfs4: do not indicate device notification if client don't expect them
fix regression introduced in 2ea7c22

On GETDEVICEINFO, client provides notify_types bitmap for mapping
notification it is interested in receiving. If the bitmap is zero
or empty, then the device ID's notifications are turned off.

We always return a non ampty bitmap independent from client's requests,
which is:

  - spec violation
  - RHEL(SL) 6 client rejects such devices

Ackek-by: Paul Millar
Target: master, 0.11