nfsv41: do not initialize back-channel
Motivation:
NFSv4.1 as a call-back mechanism which allow server to
call some procedures on the client side. We do not use
this mechanism, but ping client to check that functionality.
Unfortunately this requires, that server uses two threads -
one to process request and a second one to accept callback
reply. Drop ringing as we do not support any call-backs.

NOTICE: this is for 0.10 branch only as 0.11 and later do ping
only on MDS (door).

Modification:
remove ping over call-back channel during session initialization.

Result:
No visible user changes.

Acked-by: Gerd Behrmann
Target: 0.10