Fix leaked state resources on shutdown
Motivation:

The releasing "normal" proxy connections is bound to the disposal of the
corresponding state, which (in turn) is bound to the disposal of the
corresponding client.

When shutting down nfs4j, any existing clients are not disposed of,
resulting in no closing of any proxy connections.

Modification:

Update the NFSv4StateHandler#shutdown method to also dispose of all
clients.  While it isn't clear whether NFSv4StateHandler should be
resposible for disposing of NFS4Client objects (perhaps that should be
NFSServerV41), such a change would require an API change and this patch
is designed to be easily back-ported to supported branches.

Result:

Proxy connections are not leaked during shutdown.

Target: master
Request: 0.12
Request: 0.11
Request: 0.10
Patch: https://rb.dcache.org/r/9657/
Acked-by: Gerd Behrmann