nfs41: rename SessionSLot#checkSlotSequence to SessionSlot#acquire
Motivation:
In the order to provide serialization of the requests on the same slot,
a **acquire** + **release** semantic is planed. As **release** changes
waiting for a new functionality in under the laying rpc implementation,
we can prepare current code for up-coming change.

Modification:
rename SessionSLot#checkSlotSequence to SessionSlot#acquire

Result:
no visible changes.

Acked-by: Olufemi Adeyemi
Acked-by: Paul Millar
Target: master