nfs4: update CLOSE to use NFSv4Client#tryReleaseState
Motivation:
on close, we have to release all states, which are derived from original
open-state, like locks or layouts. However, spec requires that it CLOSE
fails if derived states can't be release.

https://tools.ietf.org/html/rfc5661#section-18.2.3

We assume, that layout is bound to open-state as well and open-state
disposal will destroy layout.

Modification:
use NFSv4Client#tryReleaseState on close to propagate such errors.
Do not call layoutreturn during close as this will be done during
state disposal. Any errors, like DELAY, will be propagated to the
client and trigger retry.

Result:
close does not directly calls layout return, but, in case of errors
propagates errors to the client.

Acked-by: Paul Millar
Target: master