vfs: change DirectoryEntry to accept cookie in the constructor
Motivation:
Directory listing can be transferred in portions. To identify which entries
are already transferred, file cookies are used. A cookie must identify files
position in the listing.

Up to now, in did, we was using list index as a cookie. This is a valid usage,
as directory listing has snapshot-like nature. When snapshot expires, BAD
COOKIE error is sent to the client. However, some nfs clients are not prepared
for such situations as they assume, that directory cookies are permanent
and if server creates a new snapshot, then cookie sill will point to the
right position.

As back-end VFS implementation may have a better way to produce cookies,
if makes sense for VFS to produce directory entries with cookies as well
as cookie-based ordered listing. A simple implementation still may use
list index as a cookies, if there are no better options available.

Modification:
Do not use list index as directory cookie. Let vfs implementation to inject
cookie to each file as it might have a better/sable way to do that. Update
READDIR implementations in v4 and v3 to relay of file-system provided cookies.

Result:
More reliable readdir operation, if vfs implementation can provide stable
cookies.

Target: master
Acked-by: Paul Millar