nfsv41: auto-configure server role as pNFS server
Motivation:
we assume that any NFSv4.1 server is a pNFS server. However this is not
the case, as the only indication of pNFS-MDS is ability to provide layouts.
A DS should not issue layouts, but have to explicitly be configured as such.

Modification:
Update CompoundContext to auto-detect server's pNFS role as following:

 - if NFSv41DeviceManager is defined - it's an MDS
 - if export file is deinfed, but not NFSv41DeviceManager, then it's
   non-pnfs server
 - if NFSv41DeviceManager and export file is not defined, then it's DS

As the server still can be misconfigured, but pNFS capability sent to
the client will reflect server's behaviour.

The OperationGETATTR is updated to advertise FATTR4_FS_LAYOUT_TYPE attribute
only if NFSv41DeviceManager is present.

Result:
We can build NFSv4.1 servers, which does not require NFSv41DeviceManager
to run in NON-PNFS or DS modes.

Acked-by: Paul Millar
Target: master