nfs4: junit: shutdown state handler after each test
Motivation:
State handler may have resources which have to be freed when not used
any more, for example an external db with client records. This have to
be done not only in production, but in testing as well.

Modification:
Update test that use NFSv4StateHandler to shutdown it after each test.
Add NFSv4StateHandler#isRunning to check the status.

Result:
No NFSv4StateHandler resource leak during tests

Acked-by: Albert Rossi
Target: master