flexfiles: make rsize/wsize configurable
Motivation:
The read/write buffer size between a and the data server is offered by
MDS. Currently, this value is hard coded to 64k, which is not optimal
value for modern hardware.

Modification:
Allow mds to specify preferred IO size for DS. Though this value can be
different for each data server, we use a single buffer size by assuming
that in typical installation data servers have a similar IO
capabilities.

As Linux client uses single value for rsize and wsize, we use (for
pragmatic reasons) a single value as well.

Result:
The IO size between clients and the server can be configured.

Acked-by: Paul Millar
Target: master