rquota: add xdrs for rquota protocol
Motivation:
Despite the fact that the NFS v4 protocol specifies attributes to query the quota
(https://tools.ietf.org/html/rfc5661#section-5.8.2.28), the Linux client does not
implement it, as the quota is not associated with a user or a group.

To support classic user and group quotas, the rquota protocol can be implemented.

Modification:
Add required xdr and server sub.

Result:
required building block to implement quota protocol.

Issue: #133
Acked-by:  Lea Morschel
Target: master