core: Make Inode a subclass of FileHandle; deprecate constructors
In an effort to simplify file handle/inode call, make Inode no longer
encapsulate FileHandle, but extend from it.

While the ultimate goal is to deprecate FileHandle, and allow Inode to
be used directly in more places, this is the first step in a series of
changes.

Deprecate the direct use of Inode constructors in favor of static helper
methods.

Adjust some places in the code that assumed encapsulation / called
constructors directly.

Related: https://github.com/dCache/nfs4j/discussions/149
Signed-off-by: Christian Kohlschütter <christian@kohlschutter.com>