basic-server: Allow for the creation of files from the outside
list() would currently fail when retrieving files in a directory that
weren't previously mapped to an Inode.

Check if the Path exists, and only throw a NoEntException if it doesn't.
In all other cases create and map a new Inode to the path.

Also handle an edge-case where a path was briefly visible but deleted
right away during list().

Related: https://github.com/dCache/nfs4j/discussions/149
Signed-off-by: Christian Kohlschütter <christian@kohlschutter.com>