nfs: drop support v0 file handles
Motivation:
as the initial nfs file handles didn't include schema version, the
FileHandle class was adopted to decode unversioned handles. The
workaround was introduced to avoid client reboot on server update.
That was in 2012...

Modification:
Drop support for v0 file handle decoding as it not needed anymore.

Result:
Less dead code.

Acked-by: Karen Hoyos
Target: master