core: test: Fix NfsServerV3READDIR_3Test, NfsServerV3READDIRPLUS_3Test
These two tests heavily use mock code and directly interact with NFS
internals.

With the combination of FileHandle and Inode, a discrepancy of expected
NFS handles was uncovered -- the NFS handle shown to clients has an
exportId of 1 (via PseudoFS), whereas the underlying handle does not.

This now creates a test error because these two Inodes are different,
and mocking code only returns a value for an exact sequence of bytes.

Change the test such that all calls to NfsServerV3.NFSPROC3_READDIR* use
the public-facing Inode (with exportId 1), and all the other code the
underlying inode (with exportId=0/unset).

Related: https://github.com/dCache/nfs4j/discussions/149
Signed-off-by: Christian Kohlschütter <christian@kohlschutter.com>