core: Opaque: Clarify mutable/immutable Opaques, add ByteBuffer support
Currently, Opaque is merely a wrapper around byte[], which adds
hashCode/equals functionality, preventing mutations, so we can use it as
a key in a HashMap.

When we process a byte stream (from a ByteBuffer), we may want to look
up values from a HashMap. However, we would still need to create new
byte[]-backed Opaque objects for the lookup. This creates unnecessary
allocations.

Improve the Opaque interface to clarify that by itself, Opaques can be
mutable. Provide a #toImmutableOpaque() method which is used when
storing values in a Map.

Adjust FileTracker to use immutable Opaques when storing in a Map.

Provide a mutable Opaque implementation backed by a ByteBuffer, and
improve equals() for OpaqueImpl.

Add a small test case to ensure sanity.

Related: https://github.com/dCache/nfs4j/discussions/149
Signed-off-by: Christian Kohlschütter <christian@kohlschutter.com>