core: Fix regression in handling of export table
As reported in #156, commit 09e8063234e4b649d3e748474d8d425f53ce64b8
caused a regression a during dCache integration.

The change threw a BadHandleException if the pseudo Inode was not the
root inode of the inner file system.

Do not throw a BadHandleException but simply convert the pseudo inode to
an inner inode.

Signed-off-by: Christian Kohlschütter <christian@kohlschutter.com>